pyproject.toml
[project]
name = "certification"
version = "0.1.0"
description = ""
authors = [
    {name = "AndyTums",email = "tumerrkin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "ipython (>=9.0.2,<10.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.2"

